requirements.txt
scikit-learn
numpy
pandas
matplotlib
pydub
librosa
tensorflow
python-dotenv